requirements.txt
pandas>=2.2.0
PyYAML>=6.0
python-dateutil>=2.8.2
chromadb>=0.4.0
qdrant-client>=1.7.0
sentence-transformers>=2.6.0
streamlit>=1.30.0
google-generativeai>=0.5.0
chromadb>=0.4.0
sentence-transformers>=2.6.0
google-genai>=0.6.0